requirements.txt
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Django==2.2.2
django-templated-mail==1.1.1
djangorestframework==3.9.4
djoser==1.7.0
ez-setup==0.9
google-api-core==1.13.0
google-auth==1.6.3
google-cloud-core==1.0.2
google-cloud-datastore==1.8.0
googleapis-common-protos==1.6.0
grpcio==1.21.1
gunicorn==19.9.0
idna==2.8
Markdown==3.1.1
protobuf==3.8.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
pytz==2019.1
requests==2.22.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
